- ssaa gui update - compatibility with plugins

+++ ACG/GL/IRenderer.hh
@@ -279,14 +279,14 @@ protected:
    * @param _outViewport [out] pointer to address of 4 GLint values receiving the currently active viewport
    * @param _outDrawBuffer [out] pointer to address receiving the currently active drawbuffer target (ie GL_BACK, GL_COLOR_ATTACHMENTi ..)
   */
-  virtual void saveActiveFbo(GLint* _outFboId, GLint* _outViewport, GLint* _outDrawBuffer) const;
+  virtual void saveActiveFbo(GLint* _outFboId, GLfloat* _outViewport, GLint* _outDrawBuffer) const;
 
   /** \brief Restore a previously saved input Fbo configuration (fbo id + viewport)
    * @param _fboId fbo id
    * @param _outViewport [in] pointer to address of 4 GLint values receiving the currently active viewport
    * @param _drawBuffer draw-buffer target of fbo
   */
-  virtual void restoreFbo(GLint _fboId, const GLint* _outViewport, GLint _drawBuffer) const;
+  virtual void restoreFbo(GLint _fboId, const GLfloat* _outViewport, GLint _drawBuffer) const;
 
   /** \brief Clear input Fbo.
    *
@@ -494,7 +494,7 @@ protected:
   GLint prevDrawBuffer_;
 
   /// previous viewport
-  GLint prevViewport_[4];
+  GLfloat prevViewport_[4];
 
   /// flag indicating a that saveCurrentFbo() has been called prior restoreFbo()
   bool prevFboSaved_;
